feat: implement add and delete product functionality in cart; add ICartItems interface

+++ src/controllers/CartsControllers.ts
@@ -1,6 +1,8 @@
 import { client } from "../db/posgres";
 import { Request, Response } from "express";
+import ICartItems from "../types/ICartItems";
 
+//shopping_carts CRUD
 export const getCarts = async (req: Request, res: Response) => {
   try {
     const response = await client.query("SELECT * FROM shopping_carts");
@@ -12,13 +14,13 @@ export const getCarts = async (req: Request, res: Response) => {
 
 export const createCart = async (req: Request, res: Response) => {
   try {
-    const userCokkiesId = req.cookies.id;
+    const userCookiesId = req.cookies.id;
 
-    if (!userCokkiesId) {
+    if (!userCookiesId) {
       res.status(401).json({ message: "Unauthorized" });
       return;
     }
-    const user_id = JSON.parse(userCokkiesId).id;
+    const user_id = JSON.parse(userCookiesId).id;
 
     const user = await client.query("SELECT * FROM users WHERE id = $1", [
       user_id,
@@ -43,6 +45,15 @@ export const createCart = async (req: Request, res: Response) => {
 export const deleteCart = async (req: Request, res: Response) => {
   try {
     const cartId = req.params.id;
+
+    const cart = await client.query("SELECT * FROM shopping_carts WHERE id = $1", [
+      cartId,
+    ]);
+    
+    if (cart.rows.length === 0) {
+      res.status(404).json({ message: "Cart not found" });
+      return;
+    }
 
     const response = await client.query(
       "DELETE FROM shopping_carts WHERE id = $1",
@@ -69,6 +80,89 @@ export const getCartById = async (req: Request, res: Response) => {
       return;
     }
     res.status(200).json({ cart: response.rows[0] });
+  } catch (error) {
+    res.status(500).json({
+      error: error,
+      message: "Internal Server Error",
+    });
+  }
+};
+
+//addProductToCart
+export const addProductToCart = async (req: Request, res: Response) => {
+  try {
+    const cartId = req.params.id;
+    const { productId, quantity }: ICartItems = req.body;
+    const cart = await client.query(
+      "SELECT * FROM shopping_carts WHERE id = $1",
+      [cartId]
+    );
+    if (cart.rows.length == 0) {
+      res.status(404).json({ message: "Cart not found" });
+      return;
+    }
+    // Check if the product already exists in the cart
+    const existingCartItem = await client.query(
+      "SELECT * FROM cart_items WHERE cart_id = $1 AND product_id = $2",
+      [cartId, productId]
+    );
+    if (existingCartItem.rows.length > 0) {
+      res.status(400).json({ message: "Product already exists in the cart" });
+      return;
+    }
+    if(quantity <= 0 || !quantity){
+      res.status(400).json({ message: "Invalid quantity" });
+      return;
+    }
+    const product = await client.query("SELECT * FROM products WHERE id = $1", [
+      productId,
+    ]);
+    if (product.rows.length == 0) {
+      res.status(404).json({ message: "Product not found" });
+      return;
+    }
+    const response = await client.query(
+      "INSERT INTO cart_items (cart_id, product_id, quantity) VALUES ($1, $2, $3) RETURNING *",
+      [cartId, productId, quantity]
+    );
+    res.status(201).json({ cart_item: response.rows[0] });
+  } catch (error) {
+    res.status(500).json({
+      error: error,
+      message: "Internal Server Error",
+    });
+  }
+};
+
+//deleteProductFromCart
+export const deleteProductFromCart = async (req: Request, res: Response) => {
+  try {
+    const cartId = req.params.id;
+    const productId = req.params.productId;
+
+    const cart = await client.query(
+      "SELECT * FROM shopping_carts WHERE id = $1",
+      [cartId]
+    );
+    if (cart.rows.length == 0) {
+      res.status(404).json({ message: "Cart not found" });
+      return;
+    }
+
+    const product = await client.query(
+      "SELECT * FROM cart_items WHERE cart_id = $1 AND product_id = $2",
+      [cartId, productId]
+    );
+    if (product.rows.length == 0) {
+      res.status(404).json({ message: "Product not found in cart" });
+      return;
+    }
+
+    await client.query(
+      "DELETE FROM cart_items WHERE cart_id = $1 AND product_id = $2",
+      [cartId, productId]
+    );
+    res.status(200).json({ message: "Product Deleted Successfully from Cart" });
   } catch (error) {
     res.status(500).json({
       error: error,
